debugging: checking for WP_error object from wp_remote_request
git-svn-id: http://plugins.svn.wordpress.org/wpbook-lite/trunk@508225 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ wpbook-lite.php
@@ -631,7 +631,13 @@ function wpbook_parse_request($wp) {
 		. '&code=' . $_REQUEST["code"];
 		// using wp_remote_request should support multiple capabilities
 		$response = wp_remote_request($token_url);
-		if( strpos($response['body'],'access_token=') !== false) {
+		if( is_wp_error($response)) {
+			echo "WP Error occured in trying to get token:\n";
+			echo "Token url was " . $token_url . "\n";
+			echo "WP Error is " . $response->get_error_message(); 
+			die(); 
+		}
+		if((is_array($response)) && (array_key_exists($response['body'])) && (strpos($response['body'],'access_token=') !== false)) {
 			// echo $response['body'];
 			$my_at = substr($response['body'],strpos($response['body'],'access_token=')+13);
 			update_option('wpbook_lite_user_access_token',$my_at);
@@ -639,6 +645,12 @@ function wpbook_parse_request($wp) {
 			echo '<a href="'. get_bloginfo('home') .'">Return to your blog</a>';
 		} else {
 			echo "Failed in creating access token"; 
+			echo "Response was ";
+			if (is_array($response)) {
+				echo print_r($response,true);
+			} else {
+				echo $response; 
+			}
 			echo '<a href="'. get_bloginfo('home') .'">Return to your blog</a>';
 		}
 		die();
